requirements.txt
Flask>=2.3
gunicorn>=21
paramiko>=3.0
scp>=0.15
Pillow>=10

